Module cyclotomic::fields::naive

Extremely simple naive implementation. Lots of copying, very inefficient, but algorithmically simple.

Structs

Number
Represents a polynomial in the orderth root of unity.

Functions

apply_automorphism

are_coprime

convert_to_base
Writes a cyclotomic in the GAP basis (keeps field the same). This is needed for elements to have a unique representation, e.g. to check equality The rewriting rule is taken from GAP but the actual code isn't.

count_powers

get_same_coeff

math_mod

one_order

phi

reduce_to_min_subfield
Expresses a cyclotomic as an element of the smallest cyclotomic field possible.

try_rational

try_reduce

zero_order