Struct cyclotomic::fields::naive::Number

pub struct Number {
    order: u64,
    coeffs: HashMap<u64, BigRational>,
}

Represents a polynomial in the orderth root of unity.

Simplest possible choice of data structure - a hash map. We assume only that each term has an exponent less than the order. So the only real term is the one for exponent zero, for example.

Fields
order: u64coeffs: HashMap<u64, BigRational>
Implementations

impl Number

pub fn new(order: &u64, coeffs: &HashMap<u64, BigRational>) -> Number

pub fn increase_order_to(z: &mut Self, new_order: &u64)

pub fn match_orders(z1: &mut Number, z2: &mut Number)